+++ CHANGELOG.md
@@ -34,6 +34,11 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Added `WPDKTerms` and `WPDKTerm` model class
 
+#### Bugs
+
+* Fixed notice/warning on shortcode `wpdk_gist` when missing attribute `file`
+
+
 
 
 ## Version 1.1.1